requirements.txt
threading
uvicorn
streamlit
pandas
streamlit.web.cli
pillow
scikit-learn